Protocol: Do not set remaining charge time if MSB is 0xff
This value is a little-endian `(u?)int16` and the most significant byte
can sometimes be set to `0xff`, which on my MY'14 car most likely means
something like "it's charging, but the remaining time is unknown". This
value occurs when I leave the charging plug connected and remotely
activate the heater. After some time, the `0xff` byte disappears and a
sensible remaining charge time appears.

This behavior reflects what `phevcore` does as well.