Protocol: Decode AC duration and state correctly
The ACMode register contains both the AC mode and the AC duration, if
set. My MY'14 would respond with the duration in the upper first bits,
causing the logic to decode the AC mode to fail.

This should be compatible with the AC mode for MY'18 cars since the
checks are bitmasked.

The AC operation state uses 5 bytes for MY'18, but only 2 for MY'14.
Since we are only interested in the second byte, broaden the check for
the data length while decoding.